Abstract out common bibliography summary
