+++ src/public/disease/sections/Bibliography/Summary.js
@@ -1,36 +1,9 @@
 import React from 'react';
 
-import { commaSeparate } from 'ot-ui';
+import BibliographySummary from '../../../common/sections/Bibliography/Summary';
 
-import { getStats } from '../../../common/sections/Bibliography/custom/Api';
-
-class Summary extends React.Component {
-  state = {
-    bibliographyCount: 0,
-  };
-  componentDidMount() {
-    const { efoId, setHasSummaryData, setHasSummaryError } = this.props;
-    getStats([{ key: efoId }]).then(
-      result => {
-        this.setState({
-          bibliographyCount: result.hits.total,
-        });
-        setHasSummaryData(result.hits.total > 0);
-      },
-      error => {
-        setHasSummaryError(true);
-      }
-    );
-  }
-  render() {
-    const { bibliographyCount } = this.state;
-    return (
-      <React.Fragment>
-        {commaSeparate(bibliographyCount)} publication
-        {bibliographyCount === 1 ? '' : 's'}
-      </React.Fragment>
-    );
-  }
-}
+const Summary = ({ efoId, ...rest }) => (
+  <BibliographySummary key={efoId} {...rest} />
+);
 
 export default Summary;
